# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SAVEONY Finance Manager API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: transactions
    description: Transaction operations
  - name: savings-goals
    description: Savings goal operations
  - name: investments
    description: Investment holding operations
  - name: dashboard
    description: Dashboard summary operations
  - name: openai
    description: OpenAI AI chat and advisor operations
  - name: budgets
    description: Budgets operations
  - name: bills
    description: Bill Reminders operations
  - name: reports
    description: Reports and historical data operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ── Transactions ─────────────────────────────────────────────────────────────
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Validation error

  /transactions/{id}:
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Savings Goals ─────────────────────────────────────────────────────────────
  /savings-goals:
    get:
      operationId: listSavingsGoals
      tags: [savings-goals]
      summary: List all savings goals
      responses:
        "200":
          description: List of savings goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavingsGoal"
    post:
      operationId: createSavingsGoal
      tags: [savings-goals]
      summary: Create a savings goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavingsGoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"
        "400":
          description: Validation error

  /savings-goals/{id}:
    patch:
      operationId: updateSavingsGoal
      tags: [savings-goals]
      summary: Update a savings goal (name, target, or current amount)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavingsGoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"
        "404":
          description: Not found
    delete:
      operationId: deleteSavingsGoal
      tags: [savings-goals]
      summary: Delete a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Investments ───────────────────────────────────────────────────────────────
  /investments:
    get:
      operationId: listInvestments
      tags: [investments]
      summary: List all investment holdings
      responses:
        "200":
          description: List of investment holdings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Investment"
    post:
      operationId: createInvestment
      tags: [investments]
      summary: Add an investment holding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
        "400":
          description: Validation error

  /investments/{id}:
    patch:
      operationId: updateInvestment
      tags: [investments]
      summary: Update an investment holding value or change
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
        "404":
          description: Not found
    delete:
      operationId: deleteInvestment
      tags: [investments]
      summary: Delete an investment holding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Dashboard Summary ─────────────────────────────────────────────────────────
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Aggregated financial summary for the dashboard
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /expenses/breakdown:
    get:
      operationId: getExpenseBreakdown
      tags: [dashboard]
      summary: Expense breakdown by category for the current month
      responses:
        "200":
          description: Expense breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseCategory"

  # ── Budgets ─────────────────────────────────────────────────────────────
  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List all budgets
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"

  /budgets/{id}:
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Bills ─────────────────────────────────────────────────────────────
  /bills:
    get:
      operationId: listBills
      tags: [bills]
      summary: List all bills
      responses:
        "200":
          description: List of bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"
    post:
      operationId: createBill
      tags: [bills]
      summary: Create a bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"

  /bills/{id}:
    patch:
      operationId: updateBill
      tags: [bills]
      summary: Update a bill (e.g., mark as paid)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
        "404":
          description: Not found
    delete:
      operationId: deleteBill
      tags: [bills]
      summary: Delete a bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Reports ─────────────────────────────────────────────────────────────
  /reports/history:
    get:
      operationId: getReportsHistory
      tags: [reports]
      summary: Get historical income and expense data
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [weekly, monthly, yearly]
      responses:
        "200":
          description: Historical data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportDataPoint"

  # ── OpenAI / AI Advisor ───────────────────────────────────────────────────────
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all AI advisor conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new AI advisor conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a message and receive a streaming AI response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    # ── Transaction ─────────────────────────────────────────────────────────────
    Transaction:
      type: object
      required: [id, merchant, category, amount, date, isIncome, createdAt]
      properties:
        id:
          type: string
        merchant:
          type: string
        category:
          type: string
        amount:
          type: number
        date:
          type: string
        isIncome:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    TransactionInput:
      type: object
      required: [merchant, category, amount, date, isIncome]
      properties:
        merchant:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        amount:
          type: number
        date:
          type: string
        isIncome:
          type: boolean
        notes:
          type: string

    # ── Savings Goal ────────────────────────────────────────────────────────────
    SavingsGoal:
      type: object
      required: [id, name, targetAmount, currentAmount, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        createdAt:
          type: string

    SavingsGoalInput:
      type: object
      required: [name, targetAmount, currentAmount]
      properties:
        name:
          type: string
          minLength: 1
        targetAmount:
          type: number
        currentAmount:
          type: number

    SavingsGoalUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        targetAmount:
          type: number
        currentAmount:
          type: number

    # ── Investment ──────────────────────────────────────────────────────────────
    Investment:
      type: object
      required: [id, ticker, name, value, allocationPercent, dayChangePercent, createdAt]
      properties:
        id:
          type: string
        ticker:
          type: string
        name:
          type: string
        value:
          type: number
        allocationPercent:
          type: number
        dayChangePercent:
          type: number
        createdAt:
          type: string

    InvestmentInput:
      type: object
      required: [ticker, name, value, allocationPercent, dayChangePercent]
      properties:
        ticker:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        value:
          type: number
        allocationPercent:
          type: number
        dayChangePercent:
          type: number

    InvestmentUpdate:
      type: object
      properties:
        ticker:
          type: string
        name:
          type: string
        value:
          type: number
        allocationPercent:
          type: number
        dayChangePercent:
          type: number

    # ── Dashboard ────────────────────────────────────────────────────────────────
    DashboardSummary:
      type: object
      required: [netWorth, monthlyIncome, monthlyExpenses, savingsRate, portfolioValue, portfolioReturnPercent, totalSaved]
      properties:
        netWorth:
          type: number
        monthlyIncome:
          type: number
        monthlyExpenses:
          type: number
        savingsRate:
          type: number
        portfolioValue:
          type: number
        portfolioReturnPercent:
          type: number
        totalSaved:
          type: number

    ExpenseCategory:
      type: object
      required: [category, total, percent]
      properties:
        category:
          type: string
        total:
          type: number
        percent:
          type: number

    # ── OpenAI / AI Advisor ───────────────────────────────────────────────────────
    OpenaiConversation:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Budgets ─────────────────────────────────────────────────────────────
    Budget:
      type: object
      required: [id, category, monthlyLimit, createdAt]
      properties:
        id:
          type: string
        category:
          type: string
        monthlyLimit:
          type: integer
        createdAt:
          type: string

    BudgetInput:
      type: object
      required: [category, monthlyLimit]
      properties:
        category:
          type: string
        monthlyLimit:
          type: integer

    # ── Bills ─────────────────────────────────────────────────────────────
    Bill:
      type: object
      required: [id, name, amount, dueDate, isPaid, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        amount:
          type: integer
        dueDate:
          type: string
        isPaid:
          type: boolean
        createdAt:
          type: string

    BillInput:
      type: object
      required: [name, amount, dueDate]
      properties:
        name:
          type: string
        amount:
          type: integer
        dueDate:
          type: string

    BillUpdate:
      type: object
      properties:
        isPaid:
          type: boolean

    # ── Reports ─────────────────────────────────────────────────────────────
    ReportDataPoint:
      type: object
      required: [date, income, expense]
      properties:
        date:
          type: string
        income:
          type: number
        expense:
          type: number